Update to Multi Switch topology
Previously, tested only with single switch star topology, basic
example. Does not test for the more complicated multi-switch topo
given in topology.in.

Unexpected behavior, of.OFPP_ALL was used instead of of.OFPP_FLOOD
causing loop of packet between switches, FLOOD prevents sending back
to switches that have sent to it.

Changed to use of.OFPP_ALL as well as to update key-to-mac store
to use Switch dpid in addition to ports to know which switch
the flow is programming. Need send both switch and port.
Since installing port is only specific to the switch.