Refactor Fault Tolerance
Previously the port entry is deleted and spanning tree is rebuilt
when TTL expires.

This may be a bit excessive and we do not know how expensive it
is for the spanning tree algo to run.

Lets make it only run the spanning tree and dump the mac_to_port
store once it detects a link is modified