Add auto removal of flow entries upon link drop
Previously, if a link drop, has to wait for the flow entries to time
out before the controller can receive any information and try to
reboot the spanning tree and receive packets from hosts that were
still on flow tables

This caused the need for shorter TTL for flows and is time dependent.
Making it such that all flows are purged once it detects links are
drop allows it to remove the time dependency and be more reactive.