Add QoS queues for premium users
Previously no QoS also bw for link was added addLink in topology.
We are unable to add queues with ovs-ofctl and queue-stats does
not detect it. We need to remove bw in linkconfig when addLink.
We will implement the max bw in queue with setting max bw on
default queue in QoS